Add a comment, rename some stuff
