+++ src/chat/tools/registerMcpTools.ts
@@ -5,7 +5,7 @@
 
 import { registerMcpToolWithTelemetry } from "@microsoft/vscode-inproc-mcp/vscode";
 import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
-import { getAzureActivityLogTool } from "./azureActivityLog/getAzureActivityLog/getAzureActivityLog";
+import { getAzureActivityLogTool } from "./azureActivityLog/getAzureActivityLog/getAzureActivityLogTool";
 import { registerCopilotOnRailsTools } from "./copilotOnRails/registerCopilotOnRailsTools";
 
 export function registerMcpTools(mcpServer: McpServer): void {
